# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexora API
  version: 1.0.0
  description: 'Nexora is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/fees/preview:
    post:
      operationId: previewFee
      summary: Preview fees
      description: Preview the calculated fees for a transfer or conversion.
      tags:
      - Fees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - type
              - asset
              - amount
              properties:
                type:
                  type: string
                  enum:
                  - transfer
                  - conversion
                asset:
                  type: string
                amount:
                  type: string
      responses:
        '200':
          description: Fee preview
          content:
            application/json:
              schema:
                type: object
                properties:
                  gross_amount:
                    type: string
                  fee_amount:
                    type: string
                  net_amount:
                    type: string
                  fee_bps:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/fees:
    get:
      operationId: getFeeSchedule
      summary: Get fee schedule
      description: Returns the fee schedule for the authenticated tenant.
      tags:
      - Fees
      responses:
        '200':
          description: Fee schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeScheduleResponse'
              example:
                transfer_fee_bps: 25
                conversion_fee_bps: 30
                min_fee_amount: '0.0100000'
                max_fee_amount: '50.0000000'
                asset: USDC
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/admin/fees/collected:
    get:
      operationId: listCollectedFees
      summary: List collected fees
      description: 'Returns fee collection totals grouped by asset and tenant.

        Requires an admin role (Owner or Admin).

        '
      tags:
      - Admin
      security:
      - AdminAuth: []
      parameters:
      - name: start_date
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: Start of period (RFC3339)
      - name: end_date
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: End of period (RFC3339)
      - name: tenant_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Filter by tenant
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
      responses:
        '200':
          description: Paginated fee collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        transaction_id:
                          type: string
                        tenant_id:
                          type: string
                        asset:
                          type: string
                        fee_amount:
                          type: string
                        fee_bps:
                          type: integer
                        collected_at:
                          type: string
                          format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
...
    Forbidden:
      description: Forbidden — admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: admin access required
              status: 403
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    FeeScheduleResponse:
      type: object
      properties:
        transfer_fee_bps:
          type: integer
          description: Transfer fee in basis points
        conversion_fee_bps:
          type: integer
          description: FX conversion fee in basis points
        min_fee_amount:
          type: string
          description: Minimum fee as decimal string
        max_fee_amount:
          type: string
          nullable: true
          description: Maximum fee cap as decimal string (null = uncapped)
        asset:
          type: string
          description: Asset in which fees are denominated
...